Refactor | Move Comment drag & drop to CommentLayer (#3150)
Move the comment drag and drop as well as the mouse listener to CommentLayer.
